pyproject.toml
[tool.poetry]
name = "rec-radiko-pg"
version = "0.1.0"
description = ""
authors = ["bugcity <wakamesukusuki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
requests = "^2.31.0"
mutagen = "^1.47.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
